refactor: move ast
