+++ tiny/src/value_object.rs
@@ -1,1 +1,2 @@
+pub mod ast;
 pub mod opcode;
